requirements.txt
flask
azure-storage-blob
azure-identity
pandas
python-dotenv
gunicorn